[DAQ-1185] use EventConstants.CMD_TOPIC instead of IEventService
each use of this generated a warning as constants should be referred
to directly. The constant is in EventConstants which IEventService
implemented. An imminent change to the scanning repo where
IEventService and EventConstants live will remove this inheritance.

also remove @Override from MockBroadcaster.setTopicName as this method
will imminently be removed from the the IPublisher interface. This is
because it is not currently called anywhere.