[DAQ-1464] update to reflect replacement of KillBean with QueueCommandBean
note: I'm not sure an instance of this class would ever receive such a
bean in the real system. Due to not sending HeartbeatBeans it wouldn't
show up in the consumer view, which allows KillBeans (now QueueCommandBeans)
to be sent.

Furthermore this class as a whole is probably not used and could be deleted.

Change-Id: I9ee00629eef1ca1ca5e74e5a7a202c240eadbbc6