[DAQ-1464] delete OperationsTestRemote and OperationSubmission
The test is the only thing that uses an OperationSubmission object,
meaning that its not used by real code. Some test methods set the queue
of a submitter after it has already been used, which doesn't make sense.
The monitor method is already commented out, and the test is never currently run.

Change-Id: I3b9769d7a361863ed86584033278cf4c3a0069e2